Foreign key implementation

+++ src/main/java/com/example/SpringBoot/repository/LoginUserDetailsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginUserDetailsRepository extends JpaRepository<LoginUserDetailsDAO, Integer> {
+public interface LoginUserDetailsRepository extends JpaRepository<LoginUserDetailsDAO, Long> {
     Optional<LoginUserDetailsDAO> findByUsername(String username);
 }
